requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
groq>=0.9.0
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
PyJWT>=2.8
passlib>=1.7.4
